Title: 세 수의 합과 평균 구하기, Time: 145ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240622/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -9,6 +9,6 @@ public class Main {
         int d = a+b+c;
         double e = (double)(a+b+c)/3;
         System.out.println(d);
-        System.out.printf("%.0f",e);
+        System.out.printf("%d",(int)e);
     }
 }
